requirements.txt
streamlit>=1.22.0
requests>=2.28.0
google-generativeai>=0.3.0
praw>=7.7.0
python-dotenv>=1.0.0 